District: fix error on delete item with offices related

+++ app/src/Backoffice/OfficeOfLearningSupportInDistrict/Application/Delete/OfficeOfLearningSupportInDistrictDeleter.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace App\Backoffice\OfficeOfLearningSupportInDistrict\Application\Delete;
 
 use App\Backoffice\OfficeOfLearningSupportInDistrict\Application\Get\Single\OfficeOfLearningSupportInDistrictFinder;
+use App\Backoffice\OfficeOfLearningSupportInDistrict\Domain\Exception\CannotDeleteDistrictWithRelatedOffices;
 use App\Backoffice\OfficeOfLearningSupportInDistrict\Domain\OfficeOfLearningSupportInDistrictRepository;
 
 final class OfficeOfLearningSupportInDistrictDeleter
@@ -22,6 +23,10 @@ final class OfficeOfLearningSupportInDistrictDeleter
     public function __invoke(string $id)
     {
         $officeOfLearningSupportInDistrict = $this->finder->__invoke($id);
+
+        if ($officeOfLearningSupportInDistrict->hasOfficesOfLearningSupport()) {
+            throw new CannotDeleteDistrictWithRelatedOffices();
+        }
 
         $this->repository->delete($officeOfLearningSupportInDistrict);
     }
